+++ .pre-commit-config.yaml
@@ -3,7 +3,7 @@ repos:
     rev: 18.9b0
     hooks:
     -   id: black
-        args: [--line-length=99, --safe]
+        args: [--safe]
         language_version: python3.7
 -   repo: https://github.com/asottile/blacken-docs
     rev: v0.3.0
